+++ src/TaskManagerKernel.java
@@ -3,7 +3,8 @@ import components.standard.Standard;
 /**
  * Task manager kernel component with primary methods.
  */
-public interface TaskManagerKernel extends Standard<TaskManager> {
+public interface TaskManagerKernel
+        extends Standard<TaskManager>, Iterable<Task> {
 
     /**
      * Adds {@code t} to {@code this}.