+++ CHANGELOG.md
@@ -52,10 +52,25 @@ setters and getters
 
 ### Added
 
-- Added removeAny to the kernel
 - Created the abstract class
-- implemented the secondary methods
+- Implemented the secondary methods
 
 ### Updated
 
+- Added removeAny to the kernel
 - Added requires and ensures claueses where necessary in method contracts
+
+## 2026.04.8
+
+### Added
+
+- Created TaskManagerOnSequence class
+- Implemented the Kernel methods using a Sequence<Task>
+
+## 2026.04.14
+
+### Updated
+
+- Implemented common methods in TaskManagerSecondary
+- Implemented common methods in Task
+- Implemented iterator for TaskManagerOnSequence